backend/requirements.txt
fastapi
uvicorn
supabase
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
gunicorn
